refactor: scope timeline dual-click affordance to empty state
