+++ src/views/timeline/timeline-panel.tsx
@@ -10,8 +10,8 @@ import { useSnapBypass } from "@/views/timeline/use-snap-bypass";
 import { useTimelineSnap } from "@/views/timeline/use-timeline-snap";
 import { ExplicitSuggestionsBanner } from "@/views/timeline/explicit-suggestions-banner";
 import { GroupingSuggestionsBanner } from "@/views/timeline/grouping-suggestions-banner";
-import { useDualClickImport } from "@/hooks/useDualClickImport";
 import { useImportModal } from "@/stores/import-modal-store";
+import { EmptyTimelineImport } from "@/views/timeline/empty-timeline-import";
 import { MarqueeSelection } from "@/views/timeline/marquee-selection";
 import { PastePreview } from "@/views/timeline/paste-preview";
 import { TimelineContextMenu } from "@/views/timeline/timeline-context-menu";
@@ -39,8 +39,7 @@ import { mainBounds } from "@/domain/line/bounds";
 import { getEffectiveLines } from "@/domain/line/effective-words";
 import { computeRowLayout, distributeLinesTiming } from "@/views/timeline/utils";
 import { GROUP_HEADER_HEIGHT } from "@/views/timeline/group-header-row";
-import { Button } from "@/ui/button";
-import { IconFileImport, IconFileMusic, IconMusic } from "@tabler/icons-react";
+import { IconMusic } from "@tabler/icons-react";
 import { DndContext, DragOverlay } from "@dnd-kit/core";
 import { useOverlayScrollbars } from "overlayscrollbars-react";
 import "overlayscrollbars/overlayscrollbars.css";
@@ -173,7 +172,6 @@ const TimelinePanel: React.FC = () => {
 
   const { marqueeRect, handleMarqueeMouseDown } = useMarquee(scrollContainerRef);
   const openLyricsModal = useCallback(() => openImportModal(), [openImportModal]);
-  const importTriggers = useDualClickImport(openLyricsModal);
   useTimelineKeyboard(scrollContainerRef, effectiveLines, duration, openLyricsModal);
   useTimelineWheel(scrollContainerRef, !!source && lines.length > 0);
 
@@ -385,23 +383,7 @@ const TimelinePanel: React.FC = () => {
     return (
       <div className="flex flex-col flex-1 overflow-hidden select-none">
         <TimelineHeader onImportLyrics={openLyricsModal} />
-        <div className="flex-1 flex flex-col items-center justify-center gap-3 p-4">
-          <IconFileMusic className="size-12 text-composer-text opacity-50" strokeWidth={1} />
-          <p className="text-lg text-composer-text-secondary">No lyrics loaded</p>
-          <p className="text-sm text-composer-text-muted">Paste lyrics or import a file</p>
-          <Button
-            variant="primary"
-            hasIcon
-            onClick={importTriggers.onClick}
-            onDoubleClick={importTriggers.onDoubleClick}
-            title="Click to search, paste, or upload. Double-click to upload a file directly."
-            className="mt-2"
-          >
-            <IconFileImport size={16} />
-            Import Lyrics
-          </Button>
-          {importTriggers.fileInput}
-        </div>
+        <EmptyTimelineImport openLyricsModal={openLyricsModal} />
       </div>
     );
   }
